Add an index.html file
Add an index.html file that's symlinked to the home.html file. This is so that when the site is live, GitHub can by default show the index.html file, which will then show links to the other pages. When the navigation bar is clicked to go back to the home.html file, it'll go there (which will look exactly like the index.html file, but with the proper URL). This is so that the first thing that users see is the home.html file, without messing with any of GitHub's routing.